+++ bedrock/bedrock-v332/src/main/java/com/nukkitx/protocol/bedrock/v332/serializer/StartGameSerializer_v332.java
@@ -16,7 +16,7 @@ public class StartGameSerializer_v332 extends StartGameSerializer_v291 { // No n
 
     @Override
     protected void writeLevelSettings(ByteBuf buffer, BedrockPacketHelper helper, StartGamePacket packet) {
-        VarInts.writeInt(buffer, packet.getSeed());
+        VarInts.writeInt(buffer, (int) packet.getSeed());
         VarInts.writeInt(buffer, packet.getDimensionId());
         VarInts.writeInt(buffer, packet.getGeneratorId());
         VarInts.writeInt(buffer, packet.getLevelGameType().ordinal());
